requirements.txt
httpx==0.24.1
python-dotenv==1.2.2
